refactory: postCategory 형변환 함수 util로 이동, when으로 형변환 함수로 변경, PostCategory&PostWritingCategory 변경
