+++ app/src/main/java/com/stopsmoke/kekkek/presentation/community/CommunityFragment.kt
@@ -1,7 +1,6 @@
 package com.stopsmoke.kekkek.presentation.community
 
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -16,7 +15,6 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.stopsmoke.kekkek.R
 import com.stopsmoke.kekkek.common.Result
-import com.stopsmoke.kekkek.core.domain.model.toPostCategory
 import com.stopsmoke.kekkek.databinding.FragmentCommunityBinding
 import com.stopsmoke.kekkek.presentation.NavigationKey
 import com.stopsmoke.kekkek.presentation.collectLatestWithLifecycle
@@ -28,6 +26,7 @@ import com.stopsmoke.kekkek.presentation.post.notice.navigateToPostNoticeScreen
 import com.stopsmoke.kekkek.presentation.post.popular.navigateToPostPopularScreen
 import com.stopsmoke.kekkek.presentation.search.navigateToSearchScreen
 import com.stopsmoke.kekkek.presentation.settings.navigateToSettingsGraph
+import com.stopsmoke.kekkek.presentation.toPostCategory
 import com.stopsmoke.kekkek.presentation.userprofile.navigateToUserProfileScreen
 import com.stopsmoke.kekkek.presentation.visible
 import dagger.hilt.android.AndroidEntryPoint
@@ -116,14 +115,16 @@ class CommunityFragment : Fragment(), ErrorHandle {
         clCommunityPostPopular1.setOnClickListener {
             val item = viewModel.uiState.value
             if (item is CommunityUiState.CommunityNormalUiState) {
-                val id = if(item.popularPeriod) item.popularItem[0].postInfo.id else item.popularItemNonPeriod[0].postInfo.id
+                val id =
+                    if (item.popularPeriod) item.popularItem[0].postInfo.id else item.popularItemNonPeriod[0].postInfo.id
                 findNavController().navigateToPostDetailScreen(id)
             }
         }
         clCommunityPostPopular2.setOnClickListener {
             val item = viewModel.uiState.value
             if (item is CommunityUiState.CommunityNormalUiState) {
-                val id = if(item.popularPeriod) item.popularItem[1].postInfo.id else item.popularItemNonPeriod[1].postInfo.id
+                val id =
+                    if (item.popularPeriod) item.popularItem[1].postInfo.id else item.popularItemNonPeriod[1].postInfo.id
                 findNavController().navigateToPostDetailScreen(id)
             }
         }
