+++ app/src/main/java/com/stopsmoke/kekkek/presentation/my/bookmark/BookmarkListViewHolder.kt
@@ -7,10 +7,10 @@ import androidx.recyclerview.widget.RecyclerView
 import coil.load
 import com.stopsmoke.kekkek.core.domain.model.DateTimeUnit
 import com.stopsmoke.kekkek.core.domain.model.ElapsedDateTime
-import com.stopsmoke.kekkek.core.domain.model.PostCategory
 import com.stopsmoke.kekkek.databinding.ItemPostBinding
 import com.stopsmoke.kekkek.presentation.community.CommunityCallbackListener
 import com.stopsmoke.kekkek.presentation.community.CommunityWritingItem
+import com.stopsmoke.kekkek.presentation.toStringKR
 
 class BookmarkListViewHolder(
     private val binding: ItemPostBinding,
@@ -38,18 +38,7 @@ class BookmarkListViewHolder(
                 ivItemWritingPostImage.visibility = View.GONE
                 setMarginEnd(tvItemWritingTitle, 16)
             }
-            tvItemWritingPostType.text =  when (item.postType) {
-                PostCategory.NOTICE -> "공지사항"
-                PostCategory.QUIT_SMOKING_SUPPORT -> "금연 지원 프로그램 공지"
-                PostCategory.POPULAR -> "인기글"
-                PostCategory.QUIT_SMOKING_AIDS_REVIEWS -> "금연 보조제 후기"
-                PostCategory.SUCCESS_STORIES -> "금연 성공 후기"
-                PostCategory.GENERAL_DISCUSSION -> "자유 게시판"
-                PostCategory.FAILURE_STORIES -> "금연 실패 후기"
-                PostCategory.RESOLUTIONS -> "금연 다짐"
-                PostCategory.UNKNOWN -> ""
-                PostCategory.ALL -> ""
-            }
+            tvItemWritingPostType.text = item.postType.toStringKR()
             tvItemWritingName.text = it.name
             tvItemWritingRank.text = "랭킹 ${it.rank}위"
