+++ app/src/main/java/com/stopsmoke/kekkek/presentation/post/notice/NoticeListAdapter.kt
@@ -14,6 +14,7 @@ import com.stopsmoke.kekkek.databinding.ItemPostBinding
 import com.stopsmoke.kekkek.presentation.community.CommunityCallbackListener
 import com.stopsmoke.kekkek.presentation.community.CommunityWritingItem
 import com.stopsmoke.kekkek.presentation.getRelativeTime
+import com.stopsmoke.kekkek.presentation.toStringKR
 import com.stopsmoke.kekkek.presentation.utils.diffutil.CommunityWritingItemDiffUtil
 
 class NoticeListAdapter
@@ -65,18 +66,7 @@ class NoticeListAdapter
                 circleIvItemWritingProfile.load(it.profileImage)
             }
 
-            tvItemWritingPostType.text = when (item.postType) {
-                PostCategory.NOTICE -> "공지사항"
-                PostCategory.QUIT_SMOKING_SUPPORT -> "금연 지원 프로그램 공지"
-                PostCategory.POPULAR -> "인기글"
-                PostCategory.QUIT_SMOKING_AIDS_REVIEWS -> "금연 보조제 후기"
-                PostCategory.SUCCESS_STORIES -> "금연 성공 후기"
-                PostCategory.GENERAL_DISCUSSION -> "자유 게시판"
-                PostCategory.FAILURE_STORIES -> "금연 실패 후기"
-                PostCategory.RESOLUTIONS -> "금연 다짐"
-                PostCategory.UNKNOWN -> ""
-                PostCategory.ALL -> ""
-            }
+            tvItemWritingPostType.text = item.postType.toStringKR()
 
             binding.root.setOnClickListener {
                 callback?.navigateToPost(item.postInfo.id)
