+++ app/src/main/java/com/stopsmoke/kekkek/presentation/post/popular/PopularPostListViewHolder.kt
@@ -6,13 +6,11 @@ import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import coil.load
 import com.stopsmoke.kekkek.R
-import com.stopsmoke.kekkek.core.domain.model.DateTimeUnit
-import com.stopsmoke.kekkek.core.domain.model.ElapsedDateTime
-import com.stopsmoke.kekkek.core.domain.model.PostCategory
 import com.stopsmoke.kekkek.databinding.ItemPostBinding
 import com.stopsmoke.kekkek.presentation.community.CommunityCallbackListener
 import com.stopsmoke.kekkek.presentation.community.CommunityWritingItem
 import com.stopsmoke.kekkek.presentation.getRelativeTime
+import com.stopsmoke.kekkek.presentation.toStringKR
 
 class PopularPostListViewHolder(
     private val binding: ItemPostBinding,
@@ -50,18 +48,7 @@ class PopularPostListViewHolder(
             }
         }
 
-        tvItemWritingPostType.text = when (item.postType) {
-            PostCategory.NOTICE -> "공지사항"
-            PostCategory.QUIT_SMOKING_SUPPORT -> "금연 지원 프로그램 공지"
-            PostCategory.POPULAR -> "인기글"
-            PostCategory.QUIT_SMOKING_AIDS_REVIEWS -> "금연 보조제 후기"
-            PostCategory.SUCCESS_STORIES -> "금연 성공 후기"
-            PostCategory.GENERAL_DISCUSSION -> "자유 게시판"
-            PostCategory.FAILURE_STORIES -> "금연 실패 후기"
-            PostCategory.RESOLUTIONS -> "금연 다짐"
-            PostCategory.UNKNOWN -> ""
-            PostCategory.ALL -> ""
-        }
+        tvItemWritingPostType.text = item.postType.toStringKR()
 
 
         binding.circleIvItemWritingProfile.setOnClickListener {
@@ -72,7 +59,6 @@ class PopularPostListViewHolder(
             callback?.navigateToPost(item.postInfo.id)
         }
     }
-
 
 
     private fun setMarginEnd(view: TextView, end: Int) {
